test(browser): Remove redundant test
